# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SlipWise expense tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List savings goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a savings goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a savings goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a savings goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── Reports / Analytics (wow endpoints) ────────────────────────────────────
  /reports/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [reports]
      summary: Dashboard summary — totals for today, week, month, and remaining budget
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          description: "Month in YYYY-MM format (defaults to current month)"
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reports/by-category:
    get:
      operationId: getSpendingByCategory
      tags: [reports]
      summary: Spending grouped by category for a period
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpend"
  /reports/by-period:
    get:
      operationId: getSpendingByPeriod
      tags: [reports]
      summary: Spending grouped by day/week/month
      security:
        - bearerAuth: []
      parameters:
        - name: granularity
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Spending over time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodSpend"
  /reports/top-merchants:
    get:
      operationId: getTopMerchants
      tags: [reports]
      summary: Top merchants by spend
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top merchants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantSpend"

  # ─── Receipt upload ────────────────────────────────────────────────────────
  /receipts/upload:
    post:
      operationId: uploadReceipt
      tags: [receipts]
      summary: Upload a receipt image and get back a URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptUpload"
      responses:
        "200":
          description: Uploaded receipt URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptUploadResult"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ─── Profile ──────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, clerkUserId, merchant, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        merchant:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Goal:
      type: object
      required: [id, clerkUserId, title, target, current, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        title:
          type: string
        target:
          type: number
        current:
          type: number
        deadline:
          type: ["string", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [title, target]
      properties:
        title:
          type: string
        target:
          type: number
        current:
          type: number
        deadline:
          type: string
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        target:
          type: number
        current:
          type: number
        deadline:
          type: string

    # ─── Reports ──────────────────────────────────────────────────────────────
    DashboardSummary:
      type: object
      required: [todaySpend, weekSpend, monthSpend, monthBudget, budgetRemaining, recentExpenses, topCategories]
      properties:
        todaySpend:
          type: number
        weekSpend:
          type: number
        monthSpend:
          type: number
        monthBudget:
          type: number
        budgetRemaining:
          type: number
        budgetPercentage:
          type: number
        recentExpenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategorySpend"
    CategorySpend:
      type: object
      required: [category, amount, count, percentage]
      properties:
        category:
          type: string
        amount:
          type: number
        count:
          type: integer
        percentage:
          type: number
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    PeriodSpend:
      type: object
      required: [period, amount, count]
      properties:
        period:
          type: string
        amount:
          type: number
        count:
          type: integer
    MerchantSpend:
      type: object
      required: [merchant, amount, count]
      properties:
        merchant:
          type: string
        amount:
          type: number
        count:
          type: integer

    # ─── Receipt ──────────────────────────────────────────────────────────────
    ReceiptUpload:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data
    ReceiptUploadResult:
      type: object
      required: [url]
      properties:
        url:
          type: string