# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SlipWise expense tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySpend } from './categorySpend';
import type { Expense } from './expense';

export interface DashboardSummary {
  todaySpend: number;
  weekSpend: number;
  monthSpend: number;
  monthBudget: number;
  budgetRemaining: number;
  budgetPercentage?: number;
  recentExpenses: Expense[];
  topCategories: CategorySpend[];
}