# lib/api-zod/src/generated/types/merchantSpend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SlipWise expense tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface MerchantSpend {
  merchant: string;
  amount: number;
  count: number;
}