# lib/api-zod/src/generated/types/receiptUpload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SlipWise expense tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ReceiptUpload {
  /** Base64-encoded image data */
  imageBase64: string;
}